+++ scripts/changelog.ts
@@ -209,6 +209,11 @@ async function exec(): Promise<void> {
         '**demo**. Log errors only',
         '**demo**. Add additional examples for plugin-usage in Storybook'
       ]
+    },
+    {
+      added: [
+        '**plugin-text**. Added Markdownshortcuts for Lists, Headings and Blockquotes.'
+      ]
     }
   ])