+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 All notable changes to this project will be documented in this file.
 
+## [Unreleased](https://github.com/edtr-io/edtr-io/compare/v0.5.0..HEAD)
+
+### Added
+
+- **plugin-text**. Added Markdownshortcuts for Lists, Headings and Blockquotes.
+
 ## [0.5.0](https://github.com/edtr-io/edtr-io/compare/v0.4.3..v0.5.0) - May 7, 2019
 
 ### Breaking Changes